requirements.txt
coinbase-advanced-py>=1.8.0
pandas>=2.0
ta>=0.11
sqlalchemy>=2.0
pyyaml>=6.0
python-dotenv>=1.0
schedule>=1.2
pytest>=8.0
pyarrow>=15.0
numpy>=1.26
